Information schema cleanup and improvements (#49)
The information schema builder requires a few small touches for the new
driver release:

- [X] It creates some tables that we don't support yet — let's remove
those queries.
- [X] It doesn't maintain the `table_constraints` table. This table will
be needed in the future, and since it also store some information about
primary and unique constraints, let's store it now, so we don't have to
backfill it with a migration later on.
- [X] Make sure identifiers are case-insensitive
- [X] Verify index and constraint naming — e.g., what happens when both
constraint name and index name are specified, and how duplicate names
are handled.