New driver fixes from manual testing (#51)
**List of discovered issues:**

- [x] Data types from `_mysql_data_types_cache` are sometimes not
correctly fetched, as they can be saved in multiple formates (lowercase,
uppercase, with or without backticks).
- [x] MySQL applies type casting when saving data in non-strict mode,
resulting in things like empty string being saved as `0` for an integer,
`0000-00-00 00:00:00` for a date, etc.
- [x] Functional defaults are incorrectly applied in non-strict mode
(`DEFAULT CURRENT_TIMESTAMP` saves a `'CURRENT_TIMESTAMP'` string rather
than the correct value).
- [x] MySQL `BINARY` type is saved as `INTEGER`, but it's actually a
binary string, and `BLOB` would be more suitable.
- [x] MySQL supports date and time components without a zero padding,
but that doesn't work with date and time functions in SQLite. (E.g.:
`2025-3-7 9:5:2` is a valid datetime/timestamp value in MySQL, but
SQLite requires it to be `2025-03-07 09:05:02`.)
- [x] In multi-process environments (php-fpm), concurrent writes can
cause "SQLSTATE[HY000]: General error: 5 database is locked"
(`SQLITE_BUSY`).