Better support for `SET` statements and MySQL variables (#214)
This pull request implements **better support for MySQL `SET`
statements**.

It also implements additional related functionality and resolves some
related issues:

- **Depth-first traversal:** All node querying APIs use depth-first
pre-order traversal. This traversal produces a natural ordering that
corresponds to the original parsed input. It also aligns with CSS
selector behavior. This is useful in situations when we need to
reconstruct or retrieve a fragment of the original input with tokens in
the correct order. It's used to preserve correct returned MySQL column
names (see the next item).
- **Correct result column names:** When no explicit column alias is
provided in a select item, the driver will now ensure that the returned
column name is equivalent to what MySQL infers from the input. For
example, if we translate `CONCAT('a', 'b')` to `('a' || 'b')`, we need
to use the original `CONCAT('a', 'b')` string as the column name. In
other words, `SELECT CONCAT('a', 'b')` will be translated to `SELECT
('a' || 'b') AS ’CONCAT('a', 'b')’`. This is also needed when selecting
user and system variables (see the items below).
- **SET NAMES/CHARSET:**`SET NAMES`, `SET CHARSET`, and `SET CHARACTER
SET` are a no-op now, with a `TODO` to implement some validation logic
in https://github.com/WordPress/sqlite-database-integration/issues/192.
- **User variables:** Reading and writing of user variables is now
implemented (`SET @my_var = ...`, `SELECT @my_var`).
- **System variables:** Reading and writing of session system variables
is now implemented (`SET time_zone = ...`, `SET @@time_zone = ...`, ,
`SET @@session.time_zone = ...`, `SELECT @@time_zone`, `SELECT
@@session.time_zone`, etc.). The variables are not yet interpreted in
any way, and they are not merged with global system variables at the
moment.

---------

Co-authored-by: Adam Zieliński <adam@adamziel.com>