Improve Query Monitor setup to better support multisite installs (#222)
This PR improves the Query Monitor setup to better support multisite
installs.

The Query Monitor integration logic tries to eagerly check whether the
Query Monitor plugin is active. This is non-trivial for multisites (at
that stage, we don't know yet which site is active, etc.).

I created a fix that doesn’t do the eager load on multisites. All is set
up correctly a bit later, when the QM plugin loads. Not loading it
eagerly means we won't log a couple of queries that WP core does before
it loads the QM plugin.

For single sites, the eager load is still in place.