Add a database name fallback for early adopters of the new driver (#226)
The experimental new SQLite driver requires a database name to be set,
while the old one doesn't.

There are some scenarios in which the `DB_NAME` constant is missing in
`wp-config.php`. To enable easier early adoption and testing before
version 3.0, we're adding a default database name in case the `DB_NAME`
constant is not set.

Since the provided database name is now verified against the existing
database, this should be a safe change.