Introduce an unsafe flag for using the new driver with SQLite < `3.37.0` (#256)
WIP open for feedback and the flag name for discussion.

I came to the conclusion that a special flag with the word `UNSAFE`
might be the best solution here.

By default, the new driver would still throw the following error:

> The SQLite version 3.27.2 is not supported. Minimum required version
is 3.37.0.

But with the new flag, it could enable accessing a database created with
3.37.0 on older SQLite versions.

**An important consideration or TODO:**
In this form, it would serve a basic use case, which is opening an
existing database on an older SQLite for previews, including writing
data to the database.

However, any new `CREATE TABLE` statement would fail because of the use
of the `STRICT` keyword (`General error: 1 near "STRICT": syntax
error`). We could address that by omitting that keyword in this mode,
but maybe let's first determine whether there are use cases for it.

Resolves
https://github.com/WordPress/sqlite-database-integration/issues/252.