Dynamic database name (#266)
This PR replaces
https://github.com/WordPress/sqlite-database-integration/pull/262. It is
a simpler solution to the same problem:

When a table reference targets an information schema table, we replace
it with a subquery, injecting the configured database name dynamically:

```sql
-- The following query:
SELECT *, t.*, t.table_schema FROM information_schema.tables t

-- Will be translated to:
SELECT *, t.*, t.table_schema FROM (
  SELECT
    `TABLE_CATALOG`,
    IIF(`TABLE_SCHEMA` = 'information_schema', `TABLE_SCHEMA`, 'database_name') AS `TABLE_SCHEMA`,
    `TABLE_NAME`,
    ...
  FROM _wp_sqlite_mysql_information_schema_tables AS tables
) t
```
The same logic will be applied to table references in JOIN clauses as
well.

---

With the new SQLite driver, we keep running into issues with missing and
incorrect database name values
(https://github.com/WordPress/sqlite-database-integration/pull/197,
https://github.com/WordPress/sqlite-database-integration/pull/203,
https://github.com/WordPress/sqlite-database-integration/pull/226,
https://github.com/WordPress/sqlite-database-integration/pull/260,
https://github.com/WordPress/sqlite-database-integration/pull/261, and
other issues).

Thinking through this further, I came to the conclusion that to provide
maximum flexibility and portability, it would be best to provide an API
in the shape of:

_"Mount `my-sqlite-file.sqlite` as `my-db-name`."_

Even if we consider adding multi-database support one day, it would
still be great to allow mounting additional SQLite files under dynamic
database names.

I don't see any downsides to doing this, except maybe that we'll have to
choose some database name in some scenarios, such as in database admin
tools. However, in these cases we can use a reasonable default or the
file name.

This is a WIP pull request demonstrating the simplest approach. For it
to be merged, I only need to resolve how the existing database name
values should be treated.