Fix "Expression tree is too large" for strings with large number of `NULL` bytes (#274)
In SQLite, it is generally difficult to work with `NULL` bytes in
strings, because SQLite uses C-style string syntax, which means it
considers the first `NULL` byte as the string terminator in some
scenarios.

While it is possible to store and read `NULL` bytes within strings (with
some limitations), there is no way to express the `NULL` byte in a
string literal. It needs to be done in a different way:

1. Concatenation: `str_part1 || CHAR(0) || str_part2 || …`
2. Passing as a parameter to a prepared statement.
3. Using HEX notation and casting to string.

More details: https://sqlite.org/nulinstr.html

At the moment, the SQLite driver uses concatenation (option 1), which
seems to fail for strings that contain many `NULL` bytes:

```
SQLSTATE[HY000]: General error: 1 Expression tree is too large (maximum depth 1000)
```

This PR fixes this by using HEX encoding.