Emulate the `VERSION()` function in the SQLite driver (#281)
Emulate the `VERSION()` function in the SQLite driver

This replaces the old implementation from the legacy SQLite translator
that's defined in `WP_SQLite_PDO_User_Defined_Functions` and returns
`5.5`.

Emulating it directly in the driver should also be faster than calling a
PHP function from SQLite.

This function is used in WordPress core and causes outdated database
warning on the Site Health screen.