Fix deprecation warnings, report them in tests (#280)
It turns out that `E_DEPRECATED` was excluded in the test bootstrap.

When enabling it, I found a few small issues, plus fixed this one:

```
PHP Deprecated:  strtolower():
Passing null to parameter #1 ($string) of type string is deprecated
in .../sqlite-database-integration/wp-includes/sqlite-ast/class-wp-sqlite-driver.php on line 2868
```

This deprecation warning occurs on PHP >= 8.4 when a SQL variable with a
`null` value is read.